requirements.txt
appnope==0.1.0
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
decorator==4.1.2
Django==1.11.4
djangorestframework==3.6.3
idna==2.5
ipython==6.1.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.10.2
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.2
requests==2.18.3
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
